split up the providers.rs
